saie-backend/requirements.txt
asgiref==3.6.0
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.26.137
botocore==1.29.137
certifi==2025.7.9
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.1
django-cleanup==9.0.0
django-cors-headers==4.0.0
django-filter==25.1
django-ses==3.5.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
gunicorn==20.1.0
h11==0.16.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
outcome==1.3.0.post0
pillow==11.2.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.4.4
tabulate==0.9.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.8.0
wsproto==1.2.0
